refactor: improve LogListFetcher by using interface and struct
Extract the mechanism for retrieving the Certificate Transparency log list into a distinct interface and HTTP-specific implementation. This improves modularity and makes the fetching process more easily testable.

Using a struct combined with an interface makes the code a lot cleaner than the previous implementation, passing a function parameter.
